requirements.txt
# cli
click
# cli.shopify.theme_updater
pyyaml
